trading/pyproject.toml
[project]
name = "lambda-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dhanhq>=2.0.2",
    "fastapi",
    "uvicorn",
    "python-dotenv>=1.2.2",
    "boto3",
    "httpx",
    "python-jose[cryptography]",
]
